Full & partian giveaways & streamline storage

+++ discord/src/main/java/pink/zak/giveawaybot/discord/pipelines/entries/steps/EligibilityCheckStep.java
@@ -33,7 +33,7 @@ public class EligibilityCheckStep {
             user.getEntries().put(messageId, Maps.newEnumMap(EntryType.class));
         }
         // no, the int is NOT a redundant cast. Please go away i know i should recode a lot of settings.
-        if (user.hasEntries(messageId) && user.getEntries(messageId).compareTo(BigInteger.valueOf(preset.getSetting(Setting.MAX_ENTRIES))) > -1) {
+        if (user.hasEntries(messageId) && user.getEntries(messageId).compareTo(BigInteger.valueOf((int) preset.getSetting(Setting.MAX_ENTRIES))) > -1) {
             return;
         }
         this.rewardStep.process(entryType, user, giveaway, preset);
